pyproject.toml
[tool.poetry]
name = "inferencedb"
version = "1.0.0"
description = ""
authors = ["Aporia <info@aporia.com>"]

[tool.poetry.dependencies]
python = ">=3.7.1 <3.11"
pydantic = "^1.9.0"
PyYAML = "^6.0"
faust = {version = "^1.10.4", extras = ["rocksdb"]}
python-schema-registry-client = "^2.2.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.12b0"
flake8 = "^4.0.1"
flake8-black = "^0.2.3"
flake8-import-order = "^0.18.1"
flake8-bugbear = "^21.11.29"
flake8-bandit = "^2.1.2"
mypy = "^0.910"
flake8-annotations = "^2.7.0"
flake8-docstrings = "^1.6.0"
darglint = "^1.8.1"
pytest-asyncio = "^0.16.0"
isort = "^5.10.1"
types-PyYAML = "^6.0.1"

[tool.poetry.scripts]
inferencedb = "inferencedb.main:main"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
force_sort_within_sections = true
lexicographical = true
order_by_type = false
group_by_package = true
no_lines_before = ['LOCALFOLDER']
line_length = 100

[tool.mypy]
show_error_codes = true
plugins = "pydantic.mypy"

[[tool.mypy.overrides]]
module = ["pandas", "pandas.*", "pytest", "apscheduler.*", "scipy.*", "uvicorn", "pyarrow.*", "sklearn.*", "parsedatetime"]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
